+++ svelte/app/components/subjectsAndTags/tags/DisplaySelectedTags.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
   import RemovableLabelDisplay from "../RemovableLabelDisplay.svelte";
+  import LabelDisplay from "../LabelDisplay.svelte";
   import { tagsState } from "app/state/tagsState";
   import FlowItems from "../../layout/FlowItems.svelte";
 
-  export let onRemove;
+  export let onRemove = null;
   export let currentlySelected;
 
   $: ({ tagHash } = $tagsState);
@@ -11,6 +12,10 @@
 
 <FlowItems tightest={true}>
   {#each currentlySelected.filter(_id => tagHash[_id]).map(_id => tagHash[_id]) as t}
-    <RemovableLabelDisplay item={t} doRemove={() => onRemove(t)} />
+    {#if onRemove}
+      <RemovableLabelDisplay item={t} doRemove={() => onRemove(t)} />
+    {:else}
+      <LabelDisplay item={t} />
+    {/if}
   {/each}
 </FlowItems>
